Add files via upload
Just make sure you create your Directory name and move into it before unzipping. I have it booting, the calculator works, showing the window and the ability to implement a toolbar. The floating menubar works, the gear icon works showing off some of the configured widgets. I'd implore you to look at how simple this is to build. Check out the apps directory. More to come as I get the server back online. Without it, the chat and web browser won't work (the web browser in the "widget maker" is an Iframe, the actual web browser window proxies through the TCP server with full browser functionality.